+++ moveit_planners/pilz_industrial_motion_planner/src/trajectory_generator_circ.cpp
@@ -43,10 +43,10 @@
 #include <kdl/utilities/error.h>
 #include <kdl/utilities/utility.h>
 #include <moveit/robot_state/conversions.hpp>
-#include <tf2_eigen_kdl/tf2_eigen_kdl.hpp>
 #include <rclcpp/logger.hpp>
 #include <rclcpp/logging.hpp>
 #include <tf2_eigen/tf2_eigen.hpp>
+#include <tf2_eigen_kdl/tf2_eigen_kdl.hpp>
 #include <tf2_geometry_msgs/tf2_geometry_msgs.hpp>
 #include <moveit/utils/logger.hpp>