+++ moveit_ros/trajectory_cache/CHANGELOG.rst
@@ -2,6 +2,11 @@
 Changelog for package moveit_ros_trajectory_cache
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+2.13.0 (2025-02-15)
+-------------------
+* Fuzzy-matching Trajectory Cache Injectable Traits refactor 🔥🔥 (`#2941 <https://github.com/ros-planning/moveit2/issues/2941>`_)
+* Contributors: methylDragon
+
 2.12.0 (2024-11-29)
 -------------------
 * Enhancement/use hpp for headers (`#3113 <https://github.com/ros-planning/moveit2/issues/3113>`_)